requirements.txt
scrapy>=2.11.0
scrapy-user-agents>=0.1.1
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.3.0
supabase>=2.0.0
